feat: add unit test coverage
